Several updates to files, adding ChEMBL matches, adding EFO ontology accessions to study file, updated bulk-map-config to inlcude ChEMBL, added processed data
