--- idr0016-wawer-30kcompounds/screenA/idr0016-screenA-bulkmap-config.yml
@@ -1,63 +1,0 @@
----
-name: idr0016-gustafsdottir-bbbc02/screenA
-version: 1
-
-defaults:
-    # Should the column be processed when creating bulk-annotations (yes/no)
-    include: no
-    # Columns type of the bulk-annotations column
-    type: string
-
-    # If non-empty a string used to separate multiple fields in a column
-    # White space will be stripped
-    split:
-    # Should this column be included in the clients (yes/no)
-    includeclient: yes
-    # Should this column be visible in the clients, if no the column should be
-    # hidden in the client but will still be indexed by the searcher (yes/no)
-    visible: yes
-    # Should empty values be omitted from the client display
-    omitempty: yes
-
-columns:
-
-  - name: Compound Broad Identifier
-    include: yes
-  - name: Compound Broad Identifier
-    clientname: Compound Broad Identifier URL
-    clientvalue: https://pubchem.ncbi.nlm.nih.gov/compound/{{ value }}
-    include: yes
-  - name: Compound Name
-    include: yes
-  - name: Compound Name
-    clientname: Compound Name URL
-    clientvalue: https://www.ncbi.nlm.nih.gov/pccompound?term={{ value }}
-    include: yes
-  - name: Phenotype 1
-    include: yes
-  - name: Phenotype 1 Term Name
-    include: yes
-  - name: Phenotype 1 Term Accession
-    include: yes
-  - name: Phenotype 1 Term Accession
-    clientname: Phenotype 1 Term Accession a URL
-    clientvalue: http://purl.obolibrary.org/obo/{{ value }}
-    include: yes
-  - name: Phenotype 2
-    include: yes
-  - name: Phenotype 2 Term Name
-    include: yes
-  - name: Phenotype 2 Term Accession
-    include: yes
-  - name: Phenotype 2 Term Accession
-    clientname: Phenotype 2 Term Accession URL
-    clientvalue: http://www.ebi.ac.uk/cmpo/{{ value }}
-    include: yes
-
-
-    
-# Advanced options (experimental)
-advanced:
-    # If a map-annotation is attached to a well also attach it to all images
-    # in the well
-    well_to_images: yes
